+++ test/support/gettext_backend.ex
@@ -8,3 +8,13 @@ defmodule TestGettext.Gettext do
     priv: "priv/gettext_test",
     plural_forms: TestBackend.Gettext.Plural
 end
+
+defmodule TestGettext.GettextUnknown do
+  @moduledoc """
+  Implements a Gettext-compatible module but using Cldr locales.  Its for
+  testing only.
+  """
+  use Gettext,
+    otp_app: Cldr.Config.app_name(),
+    priv: "priv/gettext_unknown"
+end